+++ src/main/api/fixtures/user_fixtures.py
@@ -2,8 +2,7 @@ import pytest
 
 from src.main.api.classes.api_manager import ApiManager
 from src.main.api.generators.random_model_generator import RandomModelGenerator
-from src.main.api.models.create_user_requests import CreateUserRequest
-from src.main.api.models.create_user_response import CreateUserResponse
+from src.main.api.models.create_user import CreateUserRequest, CreateUserResponse
 
 
 @pytest.fixture
