--- src/main/api/models/create_user_response.py
@@ -1,11 +1,0 @@
-from src.main.api.models.base_model import BaseModel
-from typing import Optional, List, Dict
-
-
-class CreateUserResponse(BaseModel):
-    id: int
-    username: str
-    password: str
-    name: Optional[str] = None
-    role: str
-    accounts: List[Dict[str, str]]
