--- src/main/api/models/login_user_request.py
@@ -1,6 +1,0 @@
-from src.main.api.models.base_model import BaseModel
-
-
-class LoginUserRequest(BaseModel):
-    username: str
-    password: str
