+++ src/main/api/tests/create_account_test.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.main.api.classes.api_manager import ApiManager
-from src.main.api.models.create_user_requests import CreateUserRequest
+from src.main.api.models.create_user import CreateUserRequest
 
 
 @pytest.mark.api
