+++ src/main/api/tests/create_user_test.py
@@ -3,7 +3,7 @@ import pytest
 from src.main.api.classes.api_manager import ApiManager
 from src.main.api.generators.random_data import RandomData
 from src.main.api.generators.random_model_generator import RandomModelGenerator
-from src.main.api.models.create_user_requests import CreateUserRequest
+from src.main.api.models.create_user import CreateUserRequest
 
 
 @pytest.mark.api
