+++ src/main/api/tests/login_user_test.py
@@ -1,8 +1,7 @@
 import pytest
 
 from src.main.api.classes.api_manager import ApiManager
-from src.main.api.models.create_user_requests import CreateUserRequest
-from src.main.api.models.create_user_response import CreateUserResponse
+from src.main.api.models.create_user import CreateUserRequest, CreateUserResponse
 
 
 @pytest.mark.api
